requirements.txt
Django==4.1.4
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
requests==2.28.1
